pfsetgrid failed on second execution Closes #79 (#80)
Strtok modifies the string; copy needs to be made which is then tokenized.

+++ pftools/pftools.h
@@ -70,7 +70,7 @@ typedef struct {
   Tcl_HashTable members;
   GridType grid_type;
   int total_members;
-  int num;
+  unsigned long num;
 } Data;
 
 #define DataMembers(data)   ((data)->members)
